Hackathon/requirements_txt.txt
# Core dependencies
crewai>=0.11.0
pandas>=1.5.0
matplotlib>=3.5.0
openpyxl>=3.0.0
tabulate>=0.9.0
python-dotenv>=0.19.0

# LLM providers (Groq is prioritized)
langchain-groq>=0.0.1
langchain-openai>=0.0.1
langchain-anthropic>=0.0.1
langchain-community>=0.0.1

# FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Optional for better performance
numpy>=1.24.0